api/db_setup.py
import os
from dotenv import load_dotenv
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker

path = os.path.join('..', '.env')
load_dotenv(path)
DB_USER = os.getenv("POSTGRES_USER")
DB_PASSWORD = os.getenv("POSTGRES_PASSWORD")
DB_HOST = os.getenv("POSTGRES_DB")  # Default to 'database' if not set

# Construct the SQLAlchemy Database URL
SQLALCHEMY_DATABASE_URL = f"postgresql://{DB_USER}:{DB_PASSWORD}@database:5432/{DB_HOST}"

# SQLAlchemy settings for PostgreSQL database
# SQLALCHEMY_DATABASE_URL = f"postgresql://postgres:password@database:5432/student"

# Create the SQLAlchemy engine
engine = create_engine(SQLALCHEMY_DATABASE_URL)

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
Base = declarative_base()