api/requirements.txt
fastapi[standard]
SQLAlchemy
psycopg2-binary
requests
python-dotenv